+++ R/convert_stat_chisq.R
@@ -14,8 +14,8 @@
 #' @param ci Confidence Interval (CI) level
 #' @param alternative a character string specifying the alternative hypothesis;
 #'   Controls the type of CI returned: `"greater"` (default) or `"less"`
-#'   (one-sided CI), or `"two.sided"` (default, two-sided CI). Partial matching
-#'   is allowed (e.g., `"g"`, `"l"`, `"two"`...). See *One-Sided CIs* in
+#'   (one-sided CI), or `"two.sided"` (two-sided CI). Partial matching is
+#'   allowed (e.g., `"g"`, `"l"`, `"two"`...). See *One-Sided CIs* in
 #'   [effectsize_CIs].
 #' @param adjust Should the effect size be corrected for small-sample bias?
 #'   Defaults to `TRUE`; Advisable for small samples and large tables.
@@ -48,7 +48,8 @@
 #'
 #' \deqn{\textrm{Pearson's } C = \sqrt{\chi^2 / (\chi^2 + n)}}{Pearson's C = sqrt(\chi^2 / (\chi^2 + n))}
 #'
-#' For versions adjusted for small-sample bias of \eqn{\phi}, \eqn{V}, and \eqn{T}, see [Bergsma, 2013](https://en.wikipedia.org/wiki/Cram%C3%A9r%27s_V#Bias_correction).
+#' For versions adjusted for small-sample bias of \eqn{\phi}, \eqn{V}, and \eqn{T},
+#' see [Bergsma, 2013](https://en.wikipedia.org/wiki/Cram%C3%A9r%27s_V#Bias_correction).
 #'
 #' @inheritSection effectsize_CIs Confidence (Compatibility) Intervals (CIs)
 #' @inheritSection effectsize_CIs CIs and Significance Tests