+++ R/docs_extra.R
@@ -124,13 +124,8 @@
 #' smaller than the tolerance of the optimizer, resulting in CIs of width 0.
 #' This can also result in the estimated CIs excluding the point estimate.
 #'
-#' For example:
-#' ```{r}
-#' t_to_d(80, df_error = 4555555)
-#' ```
-#'
-#' In these cases, consider an alternative optimizer, or an alternative method
-#' for computing CIs, such as the bootstrap.
+#' In these cases, consider an alternative method for computing CIs, such as the
+#' bootstrap.
 #'
 #'
 #' @references